[WINESYNC] jscript: Fix GCC 12.2 warning.
When initializing a jsstr_inline_t with a len < 3, the size passed
for the allocation is smaller then the size of the structure
(as the later is rounded up to the alignment = 4 bytes).

GCC 12.2 complains about this when dereferencing the pointer to
the structure as the size passed for allocation is smaller than the
size of the structure.

The warning is fixed by using flexible array member in
jsstr_inline_t. Given the rounding behavior in memory allocation, this
should not change the size of allocated blocks.

Signed-off-by: Eric Pouech <eric.pouech@gmail.com>

wine commit id 4bdf92cb7fa9bcd06eef318c57ae1c30fb5a2dd5 by Eric Pouech <eric.pouech@gmail.com>