[WINESYNC] rpcrt4/tests: Don't use an array of pointers in test_conf_complex_struct.
An array of pointers isn't conformant on 64-bit and will overrun the
destination buffer when unmarshalling. Since the purpose of this
function is to test a conformant FC_BOGUS_STRUCT, use one that is
conformant on both 32-bit and 64-bit.

The revised test coincidentally reveals that there is a problem with the
padding or alignment after a marshalled 16-bit enum, which will need to
be addressed separately.

Co-authored-by: Zebediah Figura <z.figura12@gmail.com>

wine commit id e09c93de7bd4f27b41f572a4a79d9df190c63c6f by Alex Henrie <alexhenrie24@gmail.com>